Split transport interface types into one file per type with DocC
Autoreview flagged TailscaleStatus.swift exporting four public types in
one file against the one-major-type-per-file rule. Move
NetworkInterfaceAddress, NetworkInterfaceAddressProviding, and
SystemNetworkInterfaceAddressProvider into their own files and document
the remaining public members.
